% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kelp_store.R
\name{object_store_kelp}
\alias{object_store_kelp}
\title{SeaweedFS-based rrq offload}
\description{
SeaweedFS-based rrq offload

SeaweedFS-based rrq offload
}
\details{
A SeaweedFS-based offload for \code{\link[rrq:object_store]{rrq::object_store}}. This is not
intended at all for direct user-use.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{object_store_kelp$new()}}
\item \href{#method-mset}{\code{object_store_kelp$mset()}}
\item \href{#method-mget}{\code{object_store_kelp$mget()}}
\item \href{#method-mdel}{\code{object_store_kelp$mdel()}}
\item \href{#method-list}{\code{object_store_kelp$list()}}
\item \href{#method-destroy}{\code{object_store_kelp$destroy()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create the store
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{object_store_kelp$new(seaweed_master_url, con, queue_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seaweed_master_url}}{URL to access seaweed master API}

\item{\code{con}}{A redis connection}

\item{\code{queue_id}}{Id of the queue, used as namespace for stored objects}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mset"></a>}}
\if{latex}{\out{\hypertarget{method-mset}{}}}
\subsection{Method \code{mset()}}{
Save a number of values to SeaweedFS
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{object_store_kelp$mset(hash, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hash}}{A character vector of object hashes}

\item{\code{value}}{A list of serialised objects
(each of which is a raw vector)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mget"></a>}}
\if{latex}{\out{\hypertarget{method-mget}{}}}
\subsection{Method \code{mget()}}{
Retrieve a number of objects from SeaweedFS
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{object_store_kelp$mget(hash)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hash}}{A character vector of hashes of the objects to return.
The objects will be deserialised before return.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mdel"></a>}}
\if{latex}{\out{\hypertarget{method-mdel}{}}}
\subsection{Method \code{mdel()}}{
Delete a number of objects from SeaweedFS
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{object_store_kelp$mdel(hash)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hash}}{A character vector of hashes to remove}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list"></a>}}
\if{latex}{\out{\hypertarget{method-list}{}}}
\subsection{Method \code{list()}}{
List hashes stored in this offload store
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{object_store_kelp$list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-destroy"></a>}}
\if{latex}{\out{\hypertarget{method-destroy}{}}}
\subsection{Method \code{destroy()}}{
Completely delete the store by removing the entire
collection from SeaweedFS and removing all data from redis.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{object_store_kelp$destroy()}\if{html}{\out{</div>}}
}

}
}
